menu: flesh out the RawFile MenuManager menu and use it in RawFileListView menu: make MenuManager a singleton() menu: move Command base class and types into its own header file to resolve circular dependency
